Add jellyfin_items_by_media_type mart, fix runtime_minutes type to numeric

+++ project-1-media-library-pipeline/media_library/models/staging/stg_jellyfin_items.sql
@@ -5,7 +5,7 @@ select
     series_name,
     nullif(year, '')::integer as year,
     genres,
-    nullif(runtime_minutes, '')::integer as runtime_minutes,
+    nullif(runtime_minutes, '')::numeric as runtime_minutes,
     nullif(file_size_bytes, '')::bigint as file_size_bytes,
     container as file_format
 from raw_jellyfin_items
